feat: implement onboarding activities and update hydration fragment UI
