--- app/src/main/java/com/example/habizen/data/OnboardingPage.kt
@@ -1,7 +1,0 @@
-package com.example.habizen.data
-
-data class OnboardingPage(
-    val title: String,
-    val description: String,
-    val illustrationRes: Int
-)
