+++ app/src/main/res/layout/fragment_hydration.xml
@@ -328,16 +328,16 @@
                         android:text="Soda\n330ml"
                         app:icon="@drawable/ic_soda" />
 
-                    <!-- Custom -->
+                    <!-- Energy Drink -->
                     <Button
-                        android:id="@+id/btnCustomDrink"
+                        android:id="@+id/btnEnergyDrink"
                         style="@style/DrinkButtonStyle"
                         android:layout_width="0dp"
                         android:layout_height="wrap_content"
                         android:layout_columnWeight="1"
                         android:layout_margin="4dp"
-                        android:text="Custom\nAmount"
-                        app:icon="@drawable/ic_add" />
+                        android:text="Energy\n200ml"
+                        app:icon="@drawable/ic_soda" />
 
                 </GridLayout>
 
@@ -718,13 +718,13 @@
 
         </androidx.cardview.widget.CardView>
 
-        <!-- Hidden Custom Amount Input (shown when Custom button clicked) -->
+        <!-- Custom Amount Input (always visible) -->
         <androidx.cardview.widget.CardView
             android:id="@+id/cardCustomAmount"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
-            android:visibility="gone"
+            android:visibility="visible"
             app:cardCornerRadius="12dp"
             app:cardElevation="4dp"
             app:cardBackgroundColor="@color/surface">
